+++ Audio/protocol.h
@@ -1,6 +1,7 @@
 #pragma once
 #include "consts.h"
 #include <array>
+#include <cstdint>
 #include <queue>
 #include <string>
 #include <vector>
@@ -13,35 +14,23 @@ std::vector<DTMF> dataToDTMF(Operation op, std::string inputData);
 std::vector<DTMF> dataToDTMF(Operation op);
 
 // Data in bits to DTMF
-std::vector<DTMF> dataBitsToDTMF(const std::vector<unsigned int> &inputData);
+std::vector<DTMF> dataBitsToDTMF(const std::vector<uint64_t> &inputData);
 
 // Translate float and string to bits
-std::vector<unsigned int> dataFloatEncode(const Operation op,
-                                          const std::vector<float> &inputData);
-std::vector<unsigned int> dataStringEncode(const Operation op,
-                                           const std::string &inputData);
+std::vector<uint64_t> dataFloatEncode(const Operation op,
+                                      const std::vector<float> &inputData);
+std::vector<uint64_t> dataStringEncode(const Operation op,
+                                       const std::string &inputData);
 
 //------------------------------------------
 // Handle sampled DTMF and present to application
-std::vector<unsigned int> DTMFtoData(std::queue<DTMF> sampleInput);
+std::vector<uint64_t> DTMFtoData(std::queue<DTMF> sampleInput);
 
 // Translate bits to operation, floats, string
-Operation getOperation(const std::vector<unsigned int> dataInput);
-std::vector<float> dataFloatDecode(std::vector<unsigned int> data);
-std::string dataStringDecode(std::vector<unsigned int> data);
+Operation getOperation(const std::vector<uint64_t> dataInput);
+std::vector<float> dataFloatDecode(std::vector<uint64_t> data);
+std::string dataStringDecode(std::vector<uint64_t> data);
 
 //-----------------------------------
-void printData(const std::vector<unsigned int> &data, int base);
-bool evenParity(const std::vector<unsigned int> &data);
-
-// Old functions
-
-/* std::vector<DTMF> dataToDTMF(Operation op, std::vector<float> inputData);
- */
-/* int dataEncode(float &inputData); */
-/* std::pair<Operation, std::vector<float>> DTMFtoData(std::deque<DTMF>
- * input);
- */
-/* std::pair<Operation, std::vector<float>> DTMFdecode(long long int &data);
- */
-/* void printData(const long long int &data, int base); */
+void printData(const std::vector<uint64_t> &data, int base);
+bool evenParity(const std::vector<uint64_t> &data);
